backend/requirements.txt
aiohttp==3.12.6
aiohappyeyeballs==2.6.1
httpx==0.28.1
httpx-sse==0.4.0
ipython==8.12.3
langchain>=0.1.0
langchain-community==0.3.24
langchain-core>=0.1.0
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph>=0.0.10
langgraph-checkpoint==2.0.26
langgraph-checkpoint-sqlite>=2.0.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
numpy==2.2.6
openai>=1.0.0
pydantic>=2.0.0
pydantic-core==2.33.2
pydantic-settings==2.9.1
python-dotenv>=1.0.0
replicate>=0.22.0
fal-client>=0.5.0
requests>=2.31.0
requests-toolbelt==1.0.0
tiktoken==0.9.0
typing-extensions==4.13.2
typing-inspect==0.9.0
PyYAML==6.0.2
zstandard==0.23.0
streamlit>=1.24.0
graphviz>=0.20.1

moviepy==2.2.1
imageio-ffmpeg>=0.4.9

# backend frontend communication
Flask
python-dotenv
Werkzeug

# Image processing for thumbnails
Pillow>=10.0.0